Release fixes 4

+++ quant_candles/signals/candles.py
@@ -1,5 +1,6 @@
 from uuid import uuid4
 
+from django.conf import settings
 from django.db.models.signals import post_delete
 from django.dispatch import receiver
 
@@ -14,4 +15,5 @@ def post_delete_candle(sender, **kwargs):
     Delete objects manually.
     """
     candle = kwargs["instance"]
-    CandleReadOnlyData.objects.filter(candle_id=candle.id).delete()
+    if settings.IS_LOCAL:
+        CandleReadOnlyData.objects.filter(candle_id=candle.id).delete()
